Add stars.db and extra stars json files, add utility to split json files, fix import utility
