--- backend/skychart/.gitignore
@@ -1,1 +1,0 @@
-stars.db
